Chapter 08 > 02. Type and 03. Subscriber Update

+++ ch08/03-subscriber/main.go
@@ -5,7 +5,7 @@ import "fmt"
 func main() {
 	var subscriber struct {
 		name     string
-		rate     float64
+		rate     int
 		isActive bool
 	}
